% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_variables.R
\docType{data}
\name{acs_variables}
\alias{acs_variables}
\title{Common American Community Survey Fields}
\format{
## `acs_variables`
A data frame with x rows and y columns:
}
\source{
<census.gov>
}
\usage{
acs_variables
}
\description{
A set of American Community Survey fields commonly used by the Child Poverty Action Lab
}
\keyword{datasets}
