% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_axis.R
\name{add_axis}
\alias{add_axis}
\title{Add Custom Axes to a [ggplot2] Plot}
\usage{
add_axis(axis = "y")
}
\arguments{
\item{axis}{A character string indicating which axis lines to add. Options are:
\itemize{
  \item `"x"`: Adds an axis line for the x-axis and removes it for the y-axis.
  \item `"y"`: Adds an axis line for the y-axis and removes it for the x-axis (default).
  \item `"both"`: Adds axis lines for both x and y axes.
}}
}
\value{
A [ggplot2] theme object modifying the axis lines.
}
\description{
Adds or removes axis lines for a specified axis in a [ggplot2] plot.
}
\examples{
\dontrun{
library(ggplot2)
ggplot(data = mtcars, aes(x = wt, y = mpg)) +
  geom_point() +
  add_axis("x")
}

}
