% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpal_plots.R
\name{add_cpal_logo}
\alias{add_cpal_logo}
\title{Add CPAL logo to plots}
\usage{
add_cpal_logo(plot, position = "bottom-right", size = 0.1, logo_path = NULL)
}
\arguments{
\item{plot}{ggplot object}

\item{position}{Position for logo: "bottom-right", "bottom-left", "top-right", "top-left"}

\item{size}{Logo size as proportion of plot (default: 0.1)}

\item{logo_path}{Path to logo file (uses package default if NULL)}
}
\value{
ggplot object with logo
}
\description{
Add the CPAL logo to a ggplot2 plot in a specified position
}
