% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpal_plots.R
\name{check_plot_accessibility}
\alias{check_plot_accessibility}
\title{Check plot accessibility}
\usage{
check_plot_accessibility(plot, verbose = TRUE)
}
\arguments{
\item{plot}{ggplot object to check}

\item{verbose}{Print detailed results (default: TRUE)}
}
\value{
List with accessibility check results
}
\description{
Performs basic accessibility checks on a ggplot object
}
