% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_web_report.R
\name{construct_web_report}
\alias{construct_web_report}
\title{construct_web_report}
\usage{
construct_web_report(name = "web_report.qmd", directory = NULL)
}
\arguments{
\item{name}{A character string for the name of the created template. Be sure
to end the name with ".Rmd".}

\item{directory}{A character string for the directory where the template
should be saved.}
}
\description{
Adds the necessary templates for making web reports at the Child Poverty Action Lab.
Adds a web report template and instructions.
}
