% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{cpal_color_pal}
\alias{cpal_color_pal}
\title{Child Poverty Action Lab [ggplot2] theme}
\usage{
cpal_color_pal(palette = "categorical")
}
\arguments{
\item{palette}{Palette name.}
}
\description{
Color palettes used at the Child Poverty Action Lab.
}
