% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{cpal_colors_extended}
\alias{cpal_colors_extended}
\title{CPAL Extended Color Palette}
\usage{
cpal_colors_extended()
}
\value{
Named character vector of extended colors
}
\description{
Extended colors including additional shades for data visualization.
All colors are loaded from `_brand.yml`.
}
\examples{
cpal_colors_extended
}
