% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpal_colors.R
\docType{data}
\name{cpal_colors_primary}
\alias{cpal_colors_primary}
\title{CPAL Primary Colors}
\format{
An object of class \code{character} of length 5.
}
\usage{
cpal_colors_primary
}
\description{
Core brand colors for CPAL visualizations
}
\keyword{datasets}
