% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{cpal_display_palettes}
\alias{cpal_display_palettes}
\title{Display CPAL Color Palettes}
\usage{
cpal_display_palettes(palette = "all")
}
\arguments{
\item{palette}{Character. Name of palette to display (or "all")}
}
\description{
Show available CPAL color palettes
}
\examples{
# Show all palettes
cpal_display_palettes("all")

# Show specific palette
cpal_display_palettes("main")
}
