% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cpal_font_family}
\alias{cpal_font_family}
\title{Get CPAL Font Family (Simple Version)}
\usage{
cpal_font_family()
}
\value{
Character string of font family name
}
\description{
Simple wrapper that returns the preferred CPAL font family.
This is a simplified version that works with gt tables.
}
