% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cpal_font_family_fallback}
\alias{cpal_font_family_fallback}
\title{CPAL Font Family Fallback}
\usage{
cpal_font_family_fallback()
}
\value{
Character string of fallback font family
}
\description{
Returns system font fallback when CPAL fonts are not available.
Used internally by theme functions.
}
