% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{cpal_geom_interactive}
\alias{cpal_geom_interactive}
\alias{cpal_point_interactive}
\alias{cpal_col_interactive}
\alias{cpal_line_interactive}
\alias{cpal_polygon_interactive}
\title{Interactive geometry layers with CPAL defaults}
\usage{
cpal_point_interactive(
  mapping = NULL,
  data = NULL,
  tooltip_var = NULL,
  onclick_var = NULL,
  data_id_var = NULL,
  ...
)

cpal_col_interactive(
  mapping = NULL,
  data = NULL,
  tooltip_var = NULL,
  onclick_var = NULL,
  data_id_var = NULL,
  ...
)

cpal_line_interactive(
  mapping = NULL,
  data = NULL,
  tooltip_var = NULL,
  onclick_var = NULL,
  data_id_var = NULL,
  ...
)

cpal_polygon_interactive(
  mapping = NULL,
  data = NULL,
  tooltip_var = NULL,
  onclick_var = NULL,
  data_id_var = NULL,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by aes()}

\item{data}{The data to be displayed in this layer}

\item{tooltip_var}{Variable to use for tooltips}

\item{onclick_var}{Variable to use for click actions}

\item{data_id_var}{Variable to use for data IDs (for linking)}

\item{...}{Other arguments passed to the geom}
}
\value{
Interactive ggplot2 layer objects with CPAL styling and ggiraph interactivity
}
\description{
These are convenience wrappers around ggiraph interactive geoms
that make it easier to add interactivity with consistent styling
}
