% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{cpal_get_primary_color}
\alias{cpal_get_primary_color}
\title{Get CPAL Primary Brand Color}
\usage{
cpal_get_primary_color()
}
\value{
A character string containing the hex color code (e.g., "#007A8C").
}
\description{
Retrieves the hex code for the primary brand color (Teal) as defined
in the `_brand.yml` palette.
}
\examples{
cpal_get_primary_color()
}
