% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpal_interactive.R
\name{cpal_interactive}
\alias{cpal_interactive}
\title{Create interactive CPAL plots with ggiraph}
\usage{
cpal_interactive(plot, width_svg = 8, height_svg = 5, ...)
}
\arguments{
\item{plot}{A ggplot2 object to make interactive}

\item{width_svg}{SVG width in inches (default: 8)}

\item{height_svg}{SVG height in inches (default: 5)}

\item{...}{Additional arguments passed to girafe()}
}
\value{
A girafe object
}
\description{
Wrapper functions to create interactive versions of ggplot2 plots
while maintaining CPAL styling
}
\examples{
\dontrun{
library(ggplot2)
library(ggiraph)

p <- ggplot(mtcars, aes(wt, mpg)) +
  geom_point_interactive(aes(tooltip = rownames(mtcars))) +
  theme_cpal()
  
cpal_interactive(p)
}
}
