% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpal_leaflet.R
\docType{data}
\name{cpal_leaflet}
\alias{cpal_leaflet}
\title{CPAL Leaflet Theme}
\format{
## `cpal_leaflet`
A single attribution row.
}
\source{
<mapbox.com>
}
\usage{
cpal_leaflet
}
\description{
Leaflet map attribution for CPAL leaflet maps.
}
\keyword{datasets}
