% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpal_interactive.R
\name{cpal_mapgl_layer}
\alias{cpal_mapgl_layer}
\title{Add CPAL-styled layer to mapgl map}
\usage{
cpal_mapgl_layer(map, id, source, type = "fill", paint = NULL, ...)
}
\arguments{
\item{map}{A mapboxgl object}

\item{id}{Layer ID}

\item{source}{Data source}

\item{type}{Layer type (fill, line, circle, etc.)}

\item{paint}{Paint properties}

\item{...}{Additional arguments}
}
\value{
Updated mapboxgl object
}
\description{
Helper function to add a layer with CPAL color defaults
}
