% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpal_note.R
\name{cpal_note}
\alias{cpal_note}
\title{cpal_note}
\usage{
cpal_note(text, size = 8, width = 132, plural = FALSE)
}
\arguments{
\item{text}{character string for a note}

\item{size}{font size for the note}

\item{width}{a number of characters to allow before a character return}

\item{plural}{If `TRUE`, will change "Note:" to "Notes:"}
}
\value{
a grob formatted for a source in a ggplot
}
\description{
Add a notes section to the bottom of a plot created with \code{cpal_plot()}.
}
