% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{cpal_palette}
\alias{cpal_palette}
\title{Get CPAL Color Palette}
\usage{
cpal_palette(palette = "primary", reverse = FALSE)
}
\arguments{
\item{palette}{Name of the palette to retrieve}

\item{reverse}{Logical. Should the palette be reversed?}
}
\value{
A vector of colors
}
\description{
Helper function to retrieve CPAL color palettes with optional reversal.
This is a wrapper around cpal_colors() for use in scale functions.
}
\keyword{internal}
