% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{cpal_palettes_categorical}
\alias{cpal_palettes_categorical}
\title{CPAL Categorical Color Palettes}
\format{
An object of class \code{list} of length 6.
}
\usage{
cpal_palettes_categorical
}
\description{
Palettes for categorical data
}
\keyword{datasets}
