% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpal_colors.R
\docType{data}
\name{cpal_palettes_diverging}
\alias{cpal_palettes_diverging}
\title{CPAL Diverging Color Palettes}
\format{
An object of class \code{list} of length 3.
}
\usage{
cpal_palettes_diverging
}
\description{
Diverging palettes for data with a meaningful midpoint
}
\keyword{datasets}
