% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpal_save.R
\name{cpal_save}
\alias{cpal_save}
\title{Save ggplot2 Plots in Standard Child Poverty Action Lab Sizes}
\usage{
cpal_save(
  filename,
  plot = ggplot2::last_plot(),
  size = "medium",
  dpi = 300,
  height = NULL
)
}
\arguments{
\item{filename}{Character string. The file name to create on disk.}

\item{plot}{A ggplot2 plot to save. Defaults to the last plot displayed.}

\item{size}{Character string. Size of the plot, must be one of `"small"` (3.25 x 2 inches),
`"medium"` (6.5 x 4 inches), or `"large"` (9 x 6.5 inches).}

\item{dpi}{Numeric or character. Plot resolution. Can be a number or `"retina"` (320), `"print"` (300), or `"screen"` (72).
Applies only to raster output types.}

\item{height}{Numeric. Custom height for the plot. Default is `NULL`, and the standard size height is used.}
}
\value{
Saves the plot to the specified file.
}
\description{
This function saves ggplot2 plots with standardized dimensions for CPAL publications.
For custom sizes or more control over width and height, consider using `ggsave()` directly.
}
\examples{
\dontrun{
library(tidyverse)
cpal_save("plot.png", size = "small")
}

}
