% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpal_interactive.R
\name{cpal_table_interactive}
\alias{cpal_table_interactive}
\title{Interactive table with reactable}
\usage{
cpal_table_interactive(data, ...)
}
\arguments{
\item{data}{Data frame to display}

\item{...}{Additional arguments passed to reactable()}
}
\value{
A reactable object
}
\description{
Create an interactive table with CPAL styling using reactable
}
