% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpal_title.R
\name{cpal_title}
\alias{cpal_title}
\title{Add a Title to a Plot}
\usage{
cpal_title(string, size = 12, width = 68)
}
\arguments{
\item{string}{A character string representing the title text.}

\item{size}{Numeric value specifying the font size for the title (default is 12).}

\item{width}{Numeric value indicating the maximum number of characters per line before wrapping the text (default is 68). If `FALSE`, no wrapping occurs.}
}
\value{
A grid graphical object (`grob`) formatted as a title for a ggplot.
}
\description{
Adds a formatted title to a plot created with \code{cpal_plot()}.
}
\examples{
\dontrun{
title <- cpal_title("This is a title")
p <- ggplot(mtcars, aes(x = wt, y = mpg)) + geom_point()
p + annotation_custom(title)
}

}
