% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{create_base_structure}
\alias{create_base_structure}
\title{Create base project directory structure}
\usage{
create_base_structure(path, type)
}
\arguments{
\item{path}{Character. Full path to project directory}

\item{type}{Character. Type of project (determines additional folders)}
}
\value{
Invisibly returns TRUE on success
}
\description{
Internal function to create standard CPAL project folders and files.
}
\keyword{internal}
