% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_bar}
\alias{geom_bar}
\title{CPAL Styled geom_bar}
\usage{
geom_bar(mapping = NULL, width = 1.1, ...)
}
\arguments{
\item{mapping}{Aesthetic mapping from ggplot2.}

\item{width}{Numeric value for the bar width (default is 1.1).}

\item{...}{Additional arguments passed to \code{geom_bar()}.}
}
\value{
A ggplot2 `geom_bar()` layer with CPAL styling.
}
\description{
This function applies a custom width and other CPAL-specific defaults to `geom_bar()`.
For full documentation, see \code{?ggplot2::geom_bar}.
}
