% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_col}
\alias{geom_col}
\title{CPAL Styled geom_col}
\usage{
geom_col(mapping = NULL, width = 1.1, ...)
}
\arguments{
\item{mapping}{Aesthetic mapping from ggplot2.}

\item{width}{Numeric value for the column width (default is 1.1).}

\item{...}{Additional arguments passed to \code{geom_col()}.}
}
\value{
A ggplot2 `geom_col()` layer with CPAL styling.
}
\description{
This function applies a custom width and other CPAL-specific defaults to `geom_col()`.
For full documentation, see \code{?ggplot2::geom_col}.
}
