% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_legend.R
\name{get_legend}
\alias{get_legend}
\title{Extract the Legend from a ggplot2 Object}
\usage{
get_legend(ggplot_object)
}
\arguments{
\item{ggplot_object}{A ggplot object from which the legend will be extracted.}
}
\value{
A grob object representing the ggplot legend, or `NULL` if no legend is present.
}
\description{
This function extracts and returns just the legend from a `ggplot2` plot object.
It's useful when you want to isolate and display the legend separately from the plot.
}
\examples{
\dontrun{
p <- ggplot(mtcars, aes(x = wt, y = mpg, color = factor(gear))) + geom_point()
legend <- get_legend(p)
grid::grid.draw(legend)
}

}
