% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{hc_theme_cpal_dark}
\alias{hc_theme_cpal_dark}
\title{CPAL Dark Theme for Highcharter}
\usage{
hc_theme_cpal_dark()
}
\value{
A Highcharter theme object
}
\description{
Returns a Highcharter theme configured with CPAL dark mode colors.
}
\examples{
\dontrun{
library(highcharter)
hchart(mtcars, "scatter", hcaes(wt, mpg)) \%>\%
  hc_add_theme(hc_theme_cpal_dark())
}
}
