% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{hc_theme_cpal_light}
\alias{hc_theme_cpal_light}
\title{CPAL Light Theme for Highcharter}
\usage{
hc_theme_cpal_light()
}
\value{
A Highcharter theme object
}
\description{
Returns a Highcharter theme configured with CPAL light mode colors.
}
\examples{
\dontrun{
library(highcharter)
hchart(mtcars, "scatter", hcaes(wt, mpg)) \%>\%
  hc_add_theme(hc_theme_cpal_light())
}
}
