% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{list_cpal_palettes}
\alias{list_cpal_palettes}
\title{List Available CPAL Palettes}
\usage{
list_cpal_palettes(details = FALSE)
}
\arguments{
\item{details}{Logical. If TRUE, show detailed information about each palette}
}
\value{
Character vector of palette names, or detailed list if details=TRUE
}
\description{
Get a list of all available CPAL color palettes with descriptions
}
\examples{
list_cpal_palettes()
list_cpal_palettes(details = TRUE)
}
