% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntx_county.R
\docType{data}
\name{ntx_county}
\alias{ntx_county}
\title{North Texas Counties}
\format{
## `ntx_county`
A list containing 8 rows and 1 column.
}
\usage{
ntx_county
}
\description{
A list containing the names of the 8 North Texas counties surrounding Dallas County.
}
\keyword{datasets}
