% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_cpal.R
\docType{data}
\name{palette_cpal}
\alias{palette_cpal}
\title{Child Poverty Action Lab [ggplot2] theme}
\format{
A \code{list}.
}
\usage{
palette_cpal
}
\description{
Palette data for the Child Poverty Action Lab
}
\details{
Data used by the palettes in the uithemes package.
}
\keyword{datasets}
