% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poppins_test.R
\name{poppins_test}
\alias{poppins_test}
\title{Test for Poppins Font Import and Registration}
\usage{
poppins_test()
}
\value{
A message indicating whether Poppins is successfully registered.
}
\description{
\code{poppins_test()} checks whether the Poppins font is imported and registered in R.
Poppins is the main font used by the Child Poverty Action Lab. If the font is not found,
instructions are provided for installing and importing it using \code{poppins_import()}.
}
\details{
You can download the font from \href{https://fonts.google.com/specimen/Poppins}{Google Fonts}.
}
