% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_legend.R
\name{remove_legend}
\alias{remove_legend}
\title{Remove the Legend from a ggplot2 Object}
\usage{
remove_legend(ggplot_object)
}
\arguments{
\item{ggplot_object}{A ggplot object from which the legend will be removed.}
}
\value{
A ggplot object without a legend.
}
\description{
This function removes the legend from a ggplot2 plot by disabling all guide aesthetics.
}
\examples{
\dontrun{
p <- ggplot(mtcars, aes(x = wt, y = mpg, color = factor(gear))) + geom_point()
p_no_legend <- remove_legend(p)
}

}
