% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_cpal.R
\name{scale_fill_cpal}
\alias{scale_fill_cpal}
\title{CPAL scale fill for ggplot2}
\usage{
scale_fill_cpal(
  theme = "factor",
  visual = list(diverge = c("#008097", "#E7ECEE", "#E98816"), triad = c("#008097",
    "#E98816", "#3f3f3f"), factor = c("#008097", "#E98816", "#3f3f3f", "#EA8B98",
    "#B5A512", "#851E2C", "#E33917"))
)
}
\description{
cpal scale fill
}
\examples{
scale_fill_cpal(theme="factor")
}
\author{
Michael Lopez
}
