% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cpal.R
\name{set_theme_cpal}
\alias{set_theme_cpal}
\title{Set CPAL Theme as Default}
\usage{
set_theme_cpal(...)
}
\arguments{
\item{...}{Arguments passed to theme_cpal()}
}
\description{
Sets theme_cpal() as the default ggplot2 theme for the session
}
