% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{setup_git}
\alias{setup_git}
\title{Setup git repository}
\usage{
setup_git(path)
}
\arguments{
\item{path}{Character. Full path to project directory}
}
\value{
Invisibly returns TRUE on success
}
\description{
Internal function to initialize git repository and create .gitignore.
}
\keyword{internal}
