% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cpal.R
\name{theme_cpal_print}
\alias{theme_cpal_print}
\title{Child Poverty Action Lab theme for print}
\usage{
theme_cpal_print(
  base_size = 8.5,
  base_family = "",
  base_line_size = 0.5,
  base_rect_size = 0.5,
  line_color = "#222222",
  rect_fill = "#E7ECEE",
  grid_color = "#E7ECEE"
)
}
\arguments{
\item{base_size}{Base font size (default is 8.5).}

\item{base_family}{Base font family (default is Inter from Google Fonts if available).}

\item{base_line_size}{Base size for lines (default is 0.5).}

\item{base_rect_size}{Base size for rectangle elements (default is 0.5).}

\item{line_color}{Default line color (default is "#222222").}

\item{rect_fill}{Background color for rectangles (default is "#E7ECEE").}

\item{grid_color}{Grid line color (default is "#E7ECEE").}
}
\description{
\code{theme_cpal_print} provides a \pkg{ggplot2} theme formatted according to the CPAL style guide for print, with sensible defaults.
}
