% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cpal.R
\name{theme_cpal_variants}
\alias{theme_cpal_variants}
\alias{theme_cpal_minimal}
\alias{theme_cpal_dark}
\alias{theme_cpal_classic}
\title{CPAL Theme Variants}
\usage{
theme_cpal_minimal(...)

theme_cpal_dark(...)

theme_cpal_classic(...)
}
\arguments{
\item{...}{Additional arguments passed to theme_cpal()}
}
\description{
Convenience functions for common theme configurations
}
