% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undo_cpal_defaults.R
\name{undo_cpal_defaults}
\alias{undo_cpal_defaults}
\title{Child Poverty Action Lab \link{ggplot2} theme}
\usage{
undo_cpal_defaults()
}
\description{
Resets all the \link{ggplot2} theme to its default settings
}
