% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_git_ignore_cpal.R
\name{use_git_ignore_cpal}
\alias{use_git_ignore_cpal}
\title{use_git_ignore_cpal}
\usage{
use_git_ignore_cpal(gitignore = "R", open = TRUE)
}
\arguments{
\item{gitignore}{Content to be added to the .gitignore file. Default is set to 'R'.}

\item{open}{Logical. If TRUE, opens the .gitignore file after creation.}
}
\description{
Adds a .gitignore file tailored for R projects.
}
