% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_readme_cpal.R
\name{use_readme_cpal}
\alias{use_readme_cpal}
\title{Add a Child Poverty Action Lab Specific README}
\usage{
use_readme_cpal(name = "README", version = "1.0", open = TRUE)
}
\arguments{
\item{name}{A character string for the name of the analysis or project (default is "README").}

\item{version}{A character string for the project or analysis version (optional).}

\item{open}{Boolean indicating whether or not to open the README file after creation (default is TRUE).}
}
\description{
\code{use_readme_cpal} adds a pre-defined CPAL README template, with the project name and optional version included at the top.
}
