% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{use_report}
\alias{use_report}
\title{Add Quarto report scaffolding to an existing project}
\usage{
use_report(path = ".", overwrite = FALSE)
}
\arguments{
\item{path}{Character; path to your project root (defaults to current working dir).}

\item{overwrite}{Logical; if TRUE, existing files will be overwritten.}
}
\value{
Invisibly, the path you supplied.
}
\description{
Copies the shared assets (CSS + images) and all report templates into
your current (or specified) project folder.
}
