% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{use_shiny}
\alias{use_shiny}
\title{Add Shiny app scaffolding to an existing project}
\usage{
use_shiny(path = ".", overwrite = FALSE)
}
\arguments{
\item{path}{Character; path to your project root (defaults to current working dir).}

\item{overwrite}{Logical; if TRUE, existing files will be overwritten.}
}
\value{
Invisibly, the path you supplied.
}
\description{
Copies the shared assets (CSS + images) and your app.R template into
your current (or specified) project folder.
}
