% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{use_shiny_dashboard}
\alias{use_shiny_dashboard}
\title{Add Shiny dashboard scaffolding to an existing project}
\usage{
use_shiny_dashboard(path = ".", overwrite = FALSE)
}
\arguments{
\item{path}{Character. Path to your project root.}

\item{overwrite}{Logical. If TRUE, overwrites existing files.}
}
\value{
Invisibly returns the project path.
}
\description{
Adds a full-featured Shiny dashboard with modules to your project.
}
