% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_web_report.R
\name{use_web_report}
\alias{use_web_report}
\title{Add a Web Report Template}
\usage{
use_web_report(name = "web_report.qmd", directory = NULL)
}
\arguments{
\item{name}{A character string for the name of the created template (default is "web_report.qmd").}

\item{directory}{A character string for the directory where the template should be saved (optional).}
}
\description{
\code{use_web_report} adds a web report template and necessary assets (such as images) to a specified directory. If a directory is provided, the template will be saved in that directory.
}
