% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{view_all_palettes}
\alias{view_all_palettes}
\title{Grid view of all palettes with hex codes}
\usage{
view_all_palettes()
}
\value{
Displays grid visualization of all color palettes (called for side effects)
}
\description{
Grid view of all palettes with hex codes
}
