% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_palette.R
\name{view_palette}
\alias{view_palette}
\title{Display a CPAL Color Palette}
\arguments{
\item{palette}{A color palette vector. Options include predefined palettes like \code{palette_cpal_diverging}, \code{palette_cpal_teal}, \code{palette_cpal_gray}, and more.}
}
\description{
\code{view_palette} displays the colors and hexadecimal codes for a specified CPAL color palette.
}
\examples{
view_palette(palette_cpal_teal)
view_palette(palette_cpal_magenta)

}
