% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{cpal_colors}
\alias{cpal_colors}
\title{Get CPAL Colors}
\usage{
cpal_colors(palette = "primary", n = NULL, reverse = FALSE)
}
\arguments{
\item{palette}{Character. Name of palette or specific colors}

\item{n}{Integer. Number of colors needed (optional)}

\item{reverse}{Logical. Reverse the order of colors?}
}
\value{
Character vector of hex colors
}
\description{
Access CPAL color palettes
}
\examples{
# Get primary colors
cpal_colors()

# Get specific colors
cpal_colors("teal")
cpal_colors(c("teal", "orange"))

# Get a sequential palette
cpal_colors("teal_seq_5")

# Get n colors from a palette
cpal_colors("main", n = 3)
}
