% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{cpal_colors_extended}
\alias{cpal_colors_extended}
\title{CPAL Extended Color Palette}
\format{
An object of class \code{character} of length 17.
}
\usage{
cpal_colors_extended
}
\description{
Extended colors including additional shades for data visualization
}
\keyword{datasets}
