% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{cpal_colors_primary}
\alias{cpal_colors_primary}
\title{CPAL Primary Colors}
\usage{
cpal_colors_primary()
}
\value{
Named character vector of primary brand colors
}
\description{
Core brand colors for CPAL visualizations. These are loaded from `_brand.yml`.
}
\examples{
cpal_colors_primary
}
