% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cpal.R
\name{cpal_font_family}
\alias{cpal_font_family}
\title{Helper function to determine available fonts}
\usage{
cpal_font_family()
}
\description{
Helper function to determine available fonts
}
\keyword{internal}
