% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cpal.R
\name{cpal_font_family_fallback}
\alias{cpal_font_family_fallback}
\title{Fallback font family detection for when Google Fonts aren't available}
\usage{
cpal_font_family_fallback()
}
\description{
Fallback font family detection for when Google Fonts aren't available
}
\keyword{internal}
