% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpal_mapbox.R
\docType{data}
\name{cpal_mapbox}
\alias{cpal_mapbox}
\title{CPAL Mapbox Theme}
\format{
## `cpal_mapbox`
A single attribution row.
}
\source{
<mapbox.com>
}
\usage{
cpal_mapbox
}
\description{
Mapbox theme link for CPAL leaflet maps.
}
\keyword{datasets}
