% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemaps.R
\docType{data}
\name{cpal_mapbox_color}
\alias{cpal_mapbox_color}
\alias{cpal_mapbox_mono}
\title{Child Poverty Action Lab Mapbox Basemaps}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\source{
Child Poverty Action Lab Data Visualization Style Guide
 \url{http://childpovertyactionlab.github.io/cpal-directory/}
}
\usage{
cpal_mapbox_color

cpal_mapbox_mono
}
\description{
Child Poverty Action Lab Mapbox Basemaps
}
\concept{cpal_mapbox}
\keyword{datasets}
