% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpal_note.R
\name{cpal_note}
\alias{cpal_note}
\title{Add Notes Section to a Plot}
\usage{
cpal_note(text, size = 8, width = 132, plural = FALSE)
}
\arguments{
\item{text}{A character string for the note to be displayed.}

\item{size}{Numeric value indicating the font size of the note (default is 8).}

\item{width}{Numeric value for the number of characters to allow before wrapping the text (default is 132).}

\item{plural}{Logical value; if `TRUE`, changes "Note:" to "Notes:" (default is `FALSE`).}
}
\value{
A grid graphical object (`grob`) formatted for a source note in a ggplot.
}
\description{
Adds a formatted notes section to the bottom of a plot created with \code{cpal_plot()}.
}
\examples{
\dontrun{
library(ggplot2)
p <- cpal_plot(mtcars, aes(x = wt, y = mpg)) + geom_point()
p + annotation_custom(cpal_note("This is a note", plural = TRUE))
}

}
