% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{palette_cpal_five}
\alias{palette_cpal_five}
\alias{palette_cpal_main}
\alias{palette_cpal_deut}
\alias{palette_cpal_diverging}
\alias{palette_cpal_quintile}
\alias{palette_cpal_teal}
\alias{palette_cpal_gray}
\alias{palette_cpal_orange}
\alias{palette_cpal_magenta}
\alias{palette_cpal_green}
\alias{palette_cpal_yellow}
\alias{palette_cpal_purple}
\title{Child Poverty Action Lab Color Palettes}
\format{
An object of class \code{character} of length 5.

An object of class \code{character} of length 8.

An object of class \code{character} of length 8.

An object of class \code{character} of length 9.

An object of class \code{character} of length 5.

An object of class \code{character} of length 8.

An object of class \code{character} of length 8.

An object of class \code{character} of length 8.

An object of class \code{character} of length 8.

An object of class \code{character} of length 8.

An object of class \code{character} of length 8.

An object of class \code{character} of length 8.
}
\source{
Child Poverty Action Lab Data Visualization Style Guide:
 \url{http://childpovertyactionlab.github.io/cpal-directory/}
}
\usage{
palette_cpal_five

palette_cpal_main

palette_cpal_deut

palette_cpal_diverging

palette_cpal_quintile

palette_cpal_teal

palette_cpal_gray

palette_cpal_orange

palette_cpal_magenta

palette_cpal_green

palette_cpal_yellow

palette_cpal_purple
}
\description{
Provides hex color codes corresponding to the official color palettes outlined in the Child Poverty Action Lab (CPAL) Data Visualization Style Guide.
These palettes are designed for consistent and accessible visualizations across CPAL projects.
}
\details{
Each palette contains a unique set of colors for different visualization needs, such as sequential, diverging, and categorical color scales. These can be used directly in plotting libraries like `ggplot2` or leaflet maps.
}
\concept{cpal_palettes}
\keyword{datasets}
