% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{palette_cpal_main}
\alias{palette_cpal_main}
\alias{palette_cpal_diverging}
\alias{palette_cpal_quintile}
\alias{palette_cpal_politics}
\alias{palette_cpal_teal}
\alias{palette_cpal_gray}
\alias{palette_cpal_orange}
\alias{palette_cpal_magenta}
\alias{palette_cpal_green}
\alias{palette_cpal_red}
\alias{palette_cpal_purple}
\title{Child Poverty Action Lab palettes}
\format{
An object of class \code{character} of length 8.

An object of class \code{character} of length 9.

An object of class \code{character} of length 5.

An object of class \code{character} of length 2.

An object of class \code{character} of length 8.

An object of class \code{character} of length 8.

An object of class \code{character} of length 8.

An object of class \code{character} of length 8.

An object of class \code{character} of length 8.

An object of class \code{character} of length 8.

An object of class \code{character} of length 8.
}
\source{
Urban Institute Data Visualization Style Guide
 \url{http://childpovertyactionlab.github.io/cpal-directory/}
}
\usage{
palette_cpal_main

palette_cpal_diverging

palette_cpal_quintile

palette_cpal_politics

palette_cpal_teal

palette_cpal_gray

palette_cpal_orange

palette_cpal_magenta

palette_cpal_green

palette_cpal_red

palette_cpal_purple
}
\description{
Vectors with hex-color codes that correspond to the color palettes outlined in the Child Poverty Action Lab Data Visualization Style Guide.
 \url{http://childpovertyactionlab.github.io/cpal-directory/}
}
\concept{cpal palettes}
\keyword{datasets}
