% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{cpal_palettes_categorical}
\alias{cpal_palettes_categorical}
\title{CPAL Categorical Color Palettes}
\usage{
cpal_palettes_categorical()
}
\value{
List of categorical color palettes
}
\description{
Palettes for categorical data, built from colors in `_brand.yml`.
Maximum 8 colors for categorical data.
}
\examples{
cpal_palettes_categorical
}
