% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{cpal_palettes_diverging}
\alias{cpal_palettes_diverging}
\title{CPAL Diverging Color Palettes}
\usage{
cpal_palettes_diverging()
}
\value{
List of diverging color palettes
}
\description{
Diverging palettes for data with a meaningful midpoint,
built from colors in `_brand.yml`.
}
\examples{
cpal_palettes_diverging
}
