% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpal_colors.R
\docType{data}
\name{cpal_palettes_sequential}
\alias{cpal_palettes_sequential}
\title{CPAL Sequential Color Palettes}
\format{
An object of class \code{list} of length 6.
}
\usage{
cpal_palettes_sequential
}
\description{
Sequential palettes for continuous data
}
\keyword{datasets}
