% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpal_plot.R
\name{cpal_plot}
\alias{cpal_plot}
\title{cpal_plot}
\usage{
cpal_plot(..., heights = 1)
}
\arguments{
\item{...}{cpal plot objects or grobs}

\item{heights}{relative heights of each object in the final plot}
}
\value{
one plot made from many grobs
}
\description{
Combine elements from \code{cpal_title}, \code{cpal_subtitle},
\code{cpal_y_title}, \code{get_legend}, \code{remove_legend},
\code{cpal_notes}, \code{cpal_source}, and \code{ggplot2} into
one formatted plot.
}
