% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpal_save.R
\name{cpal_save}
\alias{cpal_save}
\title{Save ggplot2 plots in standard Child Poverty Action Lab sizes}
\usage{
cpal_save(
  filename,
  plot = ggplot2::last_plot(),
  size = "medium",
  dpi = 300,
  height = NULL
)
}
\arguments{
\item{filename}{File name to create on disk.}

\item{plot}{Plot to save, defaults to last plot displayed.}

\item{size}{Size of plot. Must be one of "small" (3.25 x 2 inches),
"medium" (6.5 x 4 inches), or "large" (9 x 6.5 inches).}

\item{dpi}{Plot resolution. Also accepts a string input: "retina" (320),
"print" (300), or "screen" (72). Applies only to raster output types.}

\item{height}{Height of plot after save. Default is NULL.}
}
\description{
This is a function to save standardized images for CPAL
publications. For more functionality like custom widths and heights, use
ggsave() directly.
}
\examples{

\dontrun{
library(tidyverse)
library(cpalthemes)

set_cpal_defaults(style = "print")

cars \%>\%
  ggplot(aes(speed, dist)) +
  geom_point()

cpal_save("test.png")
}

}
