% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpal_source.R
\name{cpal_source}
\alias{cpal_source}
\title{cpal_source}
\usage{
cpal_source(text, size = 8, width = 132, plural = FALSE)
}
\arguments{
\item{text}{character string for a source}

\item{size}{font size for the source}

\item{width}{a number of characters to allow before a character return}

\item{plural}{If `TRUE`, will change "Source:" to "Sources:"}
}
\value{
a grob formatted for a source in a ggplot
}
\description{
Add a source section to the bottom of a plot created with \code{cpal_plot()}.
}
