% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpal_subtitle.R
\name{cpal_subtitle}
\alias{cpal_subtitle}
\title{Add a Subtitle to a Plot}
\usage{
cpal_subtitle(string, size = 9.5)
}
\arguments{
\item{string}{A character string representing the subtitle text.}

\item{size}{Numeric value specifying the font size for the subtitle (default is 9.5).}
}
\value{
A grid graphical object (`grob`) formatted for a subtitle in a ggplot.
}
\description{
Adds a formatted subtitle to a plot created with \code{cpal_plot()}.
}
\examples{
\dontrun{
subtitle <- cpal_subtitle("This is a subtitle")
p <- ggplot(mtcars, aes(x = wt, y = mpg)) + geom_point()
p + annotation_custom(subtitle)
}

}
