% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{cpal_table_gt}
\alias{cpal_table_gt}
\title{CPAL GT Table}
\usage{
cpal_table_gt(data, title = NULL, subtitle = NULL, source = NULL)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{title}{Character. The main title of the table.}

\item{subtitle}{Character. The subtitle of the table (appears below the title).}

\item{source}{Character. Source attribution text; automatically prefixed with "Source: ".}
}
\value{
A `gt_tbl` object.
}
\description{
Generates a static, publication-quality table using the `gt` framework,
styled according to CPAL brand guidelines.
}
\seealso{
[cpal_table_reactable()] for interactive versions.

Other Tables: 
\code{\link{cpal_table_reactable}()}
}
\concept{Tables}
