% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpal_title.R
\name{cpal_title}
\alias{cpal_title}
\title{cpal_title}
\usage{
cpal_title(string, size = 12, width = 68)
}
\arguments{
\item{string}{character string for a title}

\item{size}{font size for the title}

\item{width}{a number of characters to allow before a character return}
}
\value{
a grob formatted for a source in a ggplot
}
\description{
Add a title to a plot created with \code{cpal_plot()}.
}
