% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cpal.R
\name{import_inter_font}
\alias{import_inter_font}
\title{Import Inter font from Google Fonts}
\usage{
import_inter_font()
}
\value{
Logical indicating success
}
\description{
This function attempts to import the Inter font from Google Fonts
using the sysfonts and showtext packages.
}
