% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inter_import.R, R/poppins_import.R
\name{poppins_import}
\alias{poppins_import}
\title{Import and register Inter font}
\usage{
poppins_import()

poppins_import()
}
\description{
\code{inter_import()} tests to see if Inter is imported and registered. If
Inter, isn't imported and registered, then \code{inter_import()} imports and
registers Inter with
\href{https://github.com/wch/extrafont}{library(extrafont)}.

\code{poppins_import()} tests to see if Poppins is imported and registered. If
Poppins, isn't imported and registered, then \code{poppins_import()} imports and
registers Poppins with
\href{https://github.com/wch/extrafont}{library(extrafont)}.
}
\details{
Note: Inter must be installed on your computer for \code{inter_import()} to
work. Inter is the Child Poverty Action Lab's main font. To install, visit
\href{https://fonts.google.com/specimen/Poppins}{Google fonts} and click
"Download family". Unzip and open each of the .ttf files and click install.

Test to see if Inter is imported and registered with \code{inter_test()}.

Note: Poppins must be installed on your computer for \code{poppins_import()} to
work. Poppins is the Child Poverty Action Lab's main font. To install, visit
\href{https://fonts.google.com/specimen/Poppins}{Google fonts} and click
"Download family". Unzip and open each of the .ttf files and click install.

Test to see if Poppins is imported and registered with \code{poppins_test()}.
}
