% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poppins_test.R
\name{poppins_test}
\alias{poppins_test}
\title{Test for Poppins import and registration}
\usage{
poppins_test()
}
\description{
\code{poppins_test} tests to see if Poppins is imported and registered. Poppins is
the Child Poverty Action Lab's main font and can be installed from
\href{https://fonts.google.com/specimen/Poppins}{Google fonts}.
}
\details{
Import and register Poppins in R with poppins_import().
}
