% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_cpal.R
\name{scale_color_cpal}
\alias{scale_color_cpal}
\title{CPAL scale color for ggplot2}
\usage{
scale_color_cpal(
  theme = "factor",
  visual = list(diverge = c("#008097", "#E7ECEE", "#E98816"), triad = c("#008097",
    "#E98816", "#3f3f3f"), factor = c("#008097", "#E98816", "#3f3f3f", "#EA8B98",
    "#ffbe0b", "#971700", "#009763", "#6610f2"))
)
}
\arguments{
\item{theme}{Collections of colors for use in visualizations. Can be either `factor`, `triad`, or `diverge`.}

\item{visual}{List containing different color values based on theme selected.}
}
\description{
cpal scale color
}
\examples{
scale_color_cpal(theme="factor")
}
\author{
Michael Lopez
}
