% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpal_colors.R
\name{scale_cpal}
\alias{scale_cpal}
\alias{scale_color_cpal}
\alias{scale_colour_cpal}
\alias{scale_fill_cpal}
\title{CPAL Color Scales for ggplot2}
\usage{
scale_color_cpal(palette = "main", discrete = TRUE, reverse = FALSE, ...)

scale_colour_cpal(palette = "main", discrete = TRUE, reverse = FALSE, ...)

scale_fill_cpal(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character. Name of palette}

\item{discrete}{Logical. Is data discrete?}

\item{reverse}{Logical. Reverse palette?}

\item{...}{Additional arguments passed to scale functions}
}
\description{
CPAL Color Scales for ggplot2
}
