% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpal_plots.R
\name{scale_cpal_extended}
\alias{scale_cpal_extended}
\alias{scale_color_cpal_c}
\alias{scale_fill_cpal_c}
\alias{scale_color_cpal_d}
\alias{scale_fill_cpal_d}
\title{Create CPAL-branded color/fill scales}
\usage{
scale_color_cpal_c(palette = "sequential_teal", reverse = FALSE, ...)

scale_fill_cpal_c(palette = "sequential_teal", reverse = FALSE, ...)

scale_color_cpal_d(palette = "categorical", reverse = FALSE, ...)

scale_fill_cpal_d(palette = "categorical", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Palette name or custom colors}

\item{reverse}{Reverse the palette order}

\item{...}{Additional arguments passed to scale functions}
}
\description{
These functions provide additional options beyond the basic scale_color_cpal()
and scale_fill_cpal() for specific use cases.
}
