% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_grid.R
\name{scatter_grid}
\alias{scatter_grid}
\title{Child Poverty Action Lab [ggplot2] theme}
\usage{
scatter_grid()
}
\description{
Adds vertical grid lines to plots for scatter plots. This is useful for
scatter plots because cpalthemes only supplies horizontal grid lines.
}
