% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_cpal_defaults.R
\name{set_cpal_defaults}
\alias{set_cpal_defaults}
\title{Child Poverty Action Lab [ggplot2] theme}
\usage{
set_cpal_defaults(
  style = "print",
  base_size = 8.5,
  base_family = "Poppins",
  base_line_size = 0.5,
  base_rect_size = 0.5,
  scale = "continuous"
)
}
\arguments{
\item{style}{The default theme style for the R session. "print" or "map".}

\item{base_size}{The base font size for the theme. All fonts are relative to
this value.}

\item{base_family}{The base font family for the theme.}

\item{base_line_size}{The base line size for the theme. All line sizes are
relative to this value.}

\item{base_rect_size}{The base rect size for the theme. All rect sizes are
relative to this value.}

\item{scale}{For `theme_cpal_map()`. Should the legend theme be continuous or discrete?}
}
\description{
\code{set_cpal_defaults} provides a [ggplot2] theme formatted according to the
Child Poverty Action Lab style guide, with sensible defaults.
}
