% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_project.R
\name{start_project}
\alias{start_project}
\title{start_project}
\usage{
start_project(
  path,
  rstudio = rstudioapi::isAvailable(),
  open = rlang::is_interactive(),
  readme = TRUE,
  gitignore = FALSE,
  shiny = FALSE
)
}
\arguments{
\item{readme}{Add a CPAL README to the project}

\item{gitignore}{Add an R .gitignore to the project}

\item{shiny}{Adds a folder for shiny apps to the project}

\item{name}{Name of the project and the directory.}

\item{directory}{A path from the home directory of the computer to the new
directory you want created.}
}
\description{
Creates and opens a new project and project directory with useful files and folders.
}
