% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_readme_cpal.R
\name{use_readme_cpal}
\alias{use_readme_cpal}
\title{use_readme_cpal}
\usage{
use_readme_cpal(name = "README", open = TRUE)
}
\arguments{
\item{name}{Name of analysis to put at top of README}

\item{open}{Boolean for whether or not to open the instructions}
}
\description{
Adds a Child Poverty Action Lab specific README.
}
