% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{use_shiny_app}
\alias{use_shiny_app}
\title{Add simple Shiny app to an existing project}
\usage{
use_shiny_app(path = ".", overwrite = FALSE)
}
\arguments{
\item{path}{Character. Path to your project root.}

\item{overwrite}{Logical. If TRUE, overwrites existing files.}
}
\value{
Invisibly returns the project path.
}
\description{
Adds a basic Shiny application template (simpler than dashboard).
}
