% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_palette.R
\name{view_palette}
\alias{view_palette}
\title{Child Poverty Action Lab \link{ggplot2} theme}
\usage{
view_palette(palette = palette_cpal_main)
}
\arguments{
\item{palette}{A \code{palette_cpal_*} vector from \code{library(cpalthemes)}.
Options are \code{palette_cpal_diverging}, \code{palette_cpal_politics},
\code{palette_cpal_quintile}, \code{palette_cpal_teal}, \code{palette_cpal_gray},
\code{palette_cpal_yellow}, \code{palette_cpal_magenta}, \code{palette_cpal_green},
\code{palette_cpal_purple}, and \code{palette_cpal_red}.}
}
\description{
\code{view_palette} displays the colors and hexadecimal codes for \code{palette_cpal_*} vectors.
}
\examples{
view_palette()
view_palette(palette_cpal_teal)

}
