% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{view_palette}
\alias{view_palette}
\title{View CPAL Color Palettes}
\usage{
view_palette()
}
\value{
Displays color palette visualization (called for side effects)
}
\description{
Functions to visualize all available CPAL color palettes
}
\details{
Basic view of all palettes
}
